requirements.txt
python-dotenv
selenium
webdriver-manager
tqdm
bs4
lxml